groundwork for replies and characters
well, some of it. now there's sort of a system in place for stopping the
text stream when a module waits for input as well as starting it again
once that input is given.

also started implementing characters so the messages can be (visually)
distinguished based on that as well.